Enable Authelia authentication by default
The Stremio Streams Prefetcher web UI is a management/configuration
tool that should be protected with authentication. Unlike Stremio
addons which need to be publicly accessible, this tool should only
be accessible to authorized users.

Authelia middleware is already configured and running on the system,
so we enable it by default for full protection of the web interface.